% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_summaries.R
\name{eBIC}
\alias{eBIC}
\title{Calculate the extended BIC (eBIC) score of an estimated precision matrix}
\usage{
eBIC(sample.cov, theta, n, gamma = 0)
}
\arguments{
\item{sample.cov}{The sample covariance matrix of the observed data.}

\item{theta}{The estimated precision matrix.}

\item{n}{The number of observations.}

\item{gamma}{The value of the additional edge penalty parameter \eqn{\gamma} to use in the eBIC. Default value is \eqn{0}.}
}
\value{
The eBIC score.
}
\description{
Finds the extended bIC (eBIC) score of an estimated precision matrix given a sample covariance matrix, in the setting of Gaussian graphical model.
}
\examples{

# example 1: simple example where we check the eBIC score of the true precision matrix
# generate data
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
prec.mat <- dat$omega # true precision matrix
eBIC(var(dat$data), prec.mat, n, gamma = 0.2)
}
\seealso{
\code{\link{tailoredGlasso}}
}
\author{
Camilla Lingjaerde
}
