% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_summaries.R
\name{matrix.distance.simple}
\alias{matrix.distance.simple}
\title{Find the simplified matrix distance measure between two precision matrices.}
\usage{
matrix.distance.simple(mat1, mat2)
}
\arguments{
\item{mat1}{The first precision matrix. Assumed to be sparse.}

\item{mat2}{The second precision matrix. Assumed to be sparse.}
}
\value{
Numeric matrix distance.
}
\description{
Computes the matrix distance between two sparse precision matrices. The measure has a value in \eqn{[0,1]} where a value of 0 means the matrices are identical in terms of the magnitude of the corresponding partial correlations, while 1 means they are completely unrelated.
}
\examples{

# example 1: simple example with two unrelated matrices
# generate data and prior matrix
set.seed(123)
x1 <- matrix(rnorm(40 * 20), ncol = 20)
x2 <- matrix(rnorm(40 * 20), ncol = 20)
# the precision matrices
prec.mat.1 <- solve(var(x1))
prec.mat.2 <- solve(var(x2))
# make them sparse by threshold values. Do not make diagonal elements zero.
prec.mat.1[which(abs(prec.mat.1) < 1.5 & !diag(ncol(prec.mat.1)))] <- 0
prec.mat.2[which(abs(prec.mat.2) < 1.5 & !diag(ncol(prec.mat.2)))] <- 0
matrix.distance.simple(prec.mat.1, prec.mat.2)
}
\seealso{
\code{\link{tailoredGlasso}}
}
\author{
Camilla Lingjaerde
}
