% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailoredGlasso.R
\name{tailoredGlasso}
\alias{tailoredGlasso}
\title{Perform the tailored graphical lasso}
\usage{
tailoredGlasso(
  x,
  prior.matrix,
  ebic.gamma = 0,
  k.max = 100,
  stars.thresh = 0.05,
  n = NULL,
  w0 = NULL,
  k = NULL,
  lambda.opt = NULL,
  scale = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Either an \eqn{n} by \eqn{p} data matrix or a \eqn{p} by \eqn{p} empirical covariance matrix. If \code{x} is a covariance matrix, tailoredGlasso cannot perform StARS selection of \eqn{\lambda} for the unweighted grapihcal lasso graphs and \code{lambda.opt} must be provided. This is only recommended if another selection routine for \eqn{\lambda} than StARS is desired. The function automatically identifies the input matrix by checking the symmetry. \eqn{n} is the sample size and \eqn{p} is the dimension.}

\item{prior.matrix}{The \eqn{p} by \eqn{p} prior weight matrix to use. Must be symmetric and have elements in \eqn{[0,1]}.}

\item{ebic.gamma}{The value of \eqn{\gamma} to use in the extended BIC (eBIC) selection criterion. Negative values are not valid. The default value is \eqn{0}.}

\item{k.max}{The maximum value to consider for the steepness parameter \eqn{k} of the logistic function.}

\item{stars.thresh}{The variability threshold to use in the StARS selection of \eqn{\lambda} for the unweighted graph if \code{lambda.opt=NULL}.The default value is \eqn{0.05}. Will also be used to determine the sigmoid midpoint if \code{w0=NULL}.}

\item{n}{The number of observations used in the computation of the empirical covariance matrix. This quantity is used to compute the value of log-likelihood. Only required if \code{x} is an empirical covariance matrix.}

\item{w0}{If provided, the sigmoid midpoint to use for the logistic function. If \code{NULL}, it will be set to be the lower \code{stars.thresh}-quantile of the non-zero prior weights. Must be in \eqn{[0,1]}.}

\item{k}{If provided, no selection of \eqn{k} will be performed. Instead, the method will perform the weighted graphical lasso value with the prior weights transformed by the logistic function with logistic function with steepness parameter \eqn{k}. \code{NULL} by default.}

\item{lambda.opt}{If provided, the optimal \eqn{\lambda} for the unweighted graph. If not provided, it will be selected by StARS.}

\item{scale}{If \code{scale=TRUE}, all variables will be scaled. Default value is \code{FALSE}.}

\item{verbose}{If \code{verbose = FALSE}, tracing information printing is disabled. The default value is \code{TRUE}.}
}
\value{
Object of class \code{"list"}. Contains the following items:
\describe{
\item{theta.opt}{The tailored graphical lasso precision matrix estimate.}
\item{k.opt}{The selected value of \eqn{k} that minimizes the eBIC.}
\item{opt.ebic}{The eBIC value of the optimal model.}
\item{opt.sparsity}{The sparsity of the optimal model.}
\item{lambda.common}{The value of \eqn{\lambda} used in the optimal tailored graphical lasso graph.}
\item{k_vals}{All values of the steepness parameter \eqn{k} in the logistic function considered in the selection.}
\item{gamma}{The value of \eqn{\gamma} that was used in the eBIC selection.}
\item{eBIC}{The eBIC scores of the models corresponding to the different values \eqn{k}.}
\item{sparsity}{The sparsities of the models corresponding to different \eqn{k}.}
\item{loglikelihoods}{The log likelihood values of the models corresponding to different \eqn{k}.}
\item{thetas}{The estimated precision matrices corresponding to different \eqn{k}.}
\item{w0}{The sigmoid midpoint used.}
\item{opt.loglikelihood}{The log likelihood value of the selected precision matrix.}
\item{lambda.glasso}{The value of \eqn{\lambda} selected for the unweighted graphical lasso graph.}
}
}
\description{
Implements the tailored graphical lasso for data integration in Gaussian graphical models. The tailored graphical lasso utilizes useful prior information more effectively than the weighted graphical lasso without involving any risk of loss of accuracy should the prior information be misleading.
}
\details{
The tailored graphical lasso is an extension of the weighted graphical lasso for graph reconstruction. The objective is to get better utilisation of the available prior information, while ensuring that the introduction of prior information may not decrease the accuracy of the resulting inferred graph. The method takes a data matrix (or a covariance matrix) for which a weighted graphical lasso graph is to be inferred, and a prior weight matrix, and transforms the prior weights to more appropriate values with the logistic function. Unless provided, the steepness parameter \eqn{k} is chosen by the extended BIC with additional edge penalty parameter \eqn{\gamma} and the sigmoid midpoint \eqn{w_0} is chosen as the lower \eqn{\beta}-quantile of the non-zero prior weights. \eqn{\beta} is the variability threshold used in the StARS selection of the penalty parameter \eqn{\lambda} for the unweighted graphical lasso graph of the data. If \code{x} is a covariance matrix, StARS selection of \eqn{\lambda} cannot be performed and the parameter must be provided by the user.
}
\examples{

# example 1: simple example where the prior weight matrix is
# unrelated to the data of interest.
# generate data and prior matrix
set.seed(123)
x <- matrix(rnorm(40 * 20), ncol = 20)
x2 <- matrix(rnorm(40 * 20), ncol = 20)
prior.mat <- abs(cov2cor(var(x2)))
res <- tailoredGlasso(x, prior.mat)
res$theta.opt # the estimated tailoredGlasso precision matrix for x.
res$k.opt # the optimal selected value of k.
res$opt.sparsity # the sparsity of the estimated precision matrix.

# example 2: scaling the data
set.seed(123)
res <- tailoredGlasso(x, prior.mat, scale = TRUE)

# example 3: scale-free data where prior weight matrix is
# highly informative for the data of interest.
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
prec.mat <- dat$omega # true precision matrix
prior.mat <- abs(cov2cor(prec.mat))
res <- tailoredGlasso(dat$data, prior.mat, scale = TRUE)
res$k.opt
precision(abs(prec.mat) > 1e-7, res$theta.opt != 0)
# k is chosen very large, high precision.

# example 4: scale-free data where prior weight matrix is
#  completely uninformative for the data of interest.
# Create a completely unrealted prior data set
set.seed(123)
n <- 80
p <- 100
dat <- huge::huge.generator(n = n, d = p, graph = "scale-free")
dat.prior <- huge::huge.generator(n = n, d = p, graph = "scale-free")
prec.mat.prior <- dat.prior$omega # true precision matrix
prior.mat <- abs(cov2cor(prec.mat.prior))
res <- tailoredGlasso(dat$data, prior.mat, scale = TRUE)
res$k.opt
precision(abs(dat$omega) > 1e-7, res$theta.opt != 0)
# k is chosen to be very small, lower precision as less useful prior information.
}
\seealso{
\code{\link[huge]{huge}}, \code{\link[huge]{huge.select}}, \code{\link[glasso]{glasso}}
}
\author{
Camilla Lingjaerde
}
